splitting out vault detail
