+++ tauri-app/src/lib/components/charts/LightweightChart.svelte
