+++ tauri-app/src/lib/components/charts/LightweightChartHistogram.svelte
