+++ tauri-app/src/lib/components/charts/LightweightChartLine.svelte
