+++ tauri-app/src/lib/components/detail/TanstackPageContentDetail.svelte
