+++ tauri-app/src/lib/components/detail/TanstackPageContentDetail.test.ts
